Fix SOCKS5 byte alignment
This reverts the change to src/tcp/socks5/Socks5.hpp made by 17ca882 as it
prevented SOCKS5 communication: all connections would fail with EOF.

+++ src/tcp/socks5/Socks5.hpp
@@ -40,7 +40,7 @@ class Socks5
   struct initRequest
   {
     const uint8_t version = 0x05;
-    size_t nmethods;
+    uint8_t nmethods;
     AuthMethod methods[255];
   } initRequest_;
